requirements.txt
gradio>=4.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
torch>=2.0.0
numpy>=1.24.0
openai-whisper>=20231117
bitsandbytes>=0.41.0
accelerate>=0.21.0
google-generativeai>=0.3.0
